Fix broken code and warnings
* Add missing include.

* Ensure that strict-aliasing can be enforced with teh CAS operation. A
memcpy is used in place of a reinterpret_cast.

+++ core/bigvector.hpp
@@ -68,7 +68,7 @@ public:
 			FILE * fout = fopen(path.c_str(), "wb");
 			fclose(fout);
 		}
-		if (file_size(path) != sizeof(T) * length) {
+		if (static_cast<size_t>(file_size(path)) != sizeof(T) * length) {
 			long file_length = sizeof(T) * length;
 			assert(truncate(path.c_str(), file_length)!=-1);
 			int fout = open(path.c_str(), O_WRONLY);
